requirements.txt
streamlit
python-docx
scikit-learn
nltk
spacy
langchain
openai
python-dotenv
numpy
pandas